+++ extension.js
@@ -7,7 +7,14 @@ const vscode = require("vscode");
 function activate(context) {
 	console.log('Congratulations, your extension "code-typewriter-effect" is now active!');
 
-	let disposable = vscode.commands.registerCommand("code-typewriter-effect.retype", function () {
+	let typingProcessRunning = false; // Variable to track if the typing process is running
+
+	let disposableType = vscode.commands.registerCommand("code-typewriter-effect.retype", function () {
+		if (typingProcessRunning) {
+			return; // If the typing process is already running, do nothing
+		}
+
+		typingProcessRunning = true; // Mark that the typing process has started
 
 		let editor = vscode.window.activeTextEditor;
 		if (!editor) {
@@ -24,10 +31,18 @@ function activate(context) {
 		});
 
 		function type() {
+			if (!typingProcessRunning) {
+				return; // If the typing process is stopped, exit the function
+			}
+
 			let line = text.split("\n")[currentPosition];
 			let i = 0;
 
 			function insertNextCharacter() {
+				if (!typingProcessRunning) {
+					return; // If the typing process is stopped, exit the function
+				}
+
 				if (i <= line.length) {
 					editor.edit((editBuilder) => {
 						editBuilder.insert(new vscode.Position(currentPosition, i), line.charAt(i));
@@ -41,6 +56,8 @@ function activate(context) {
 						line = text.split("\n")[currentPosition];
 						i = 0;
 						setTimeout(insertNextCharacter, 0);
+					} else {
+						typingProcessRunning = false; // Typing process is complete, reset the variable
 					}
 				}
 			}
@@ -49,7 +66,11 @@ function activate(context) {
 		}
 	});
 
-	context.subscriptions.push(disposable);
+	let disposableStop = vscode.commands.registerCommand("code-typewriter-effect.stop", function () {
+		typingProcessRunning = false; // When the "code-typewriter-effect.stop" command is executed, stop the typing process
+	});
+
+	context.subscriptions.push(disposableType, disposableStop);
 }
 
 function deactivate() { }